Refactoring Manager.addPlayer() method.
In this branch I'll try to forward arguments for Player ctor in
addPlayer method. This is not good solution, IMHO.
According to this, list of argumants for ctor requre to be
'canonized', whitch binds hands for other defelopers.
On the other hand, Orleg says, that it's fine. Than this should
be done in another way. Yet another time -- IMHO.