Adding RrnderMachine class and RenderInterface
This invokes declaration of RenderInterface::draw() method, whitch
uses a bit different policies.
The draw() methods should be implemented in a separated file with
the other stuff, needed for the choosen graphic library, or method.
The other stuff -- it is the RenderMachine methods, whitch used for
all interactins with user.