Adding SfmlRender class
Here, I chosed the SFML graphic library. To be honest, the most of
things are already exist in the SFML. But I don't care. The only
one reason, why I used it, is the documentation.
Well... not the only one, but definitely not the last one.
Maybe, OpenGL would be better in this situation...
If Orleg would be upset due to I implementing the graphics -- I'll
let him to use anything ehat he wants. Flexibility, damn it)